Move actions to the correct place and ease usage of parameters
